refactor: 필요없는 CouponServiceDelegator 삭제
Related to: #113
