+++ src/main/java/com/hf/healthfriend/domain/coupon/dto/request/GrantCouponDto.java
@@ -1,5 +1,6 @@
 package com.hf.healthfriend.domain.coupon.dto.request;
 
+import com.hf.healthfriend.domain.coupon.entity.Coupon;
 import lombok.*;
 
 import java.time.temporal.ChronoUnit;
@@ -12,4 +13,6 @@ public abstract class GrantCouponDto {
     private Long receiverId;
     private long validTimeAmount;
     private ChronoUnit validTimeUnit;
+
+    public abstract Coupon toEntity();
 }
