+++ src/main/java/com/hf/healthfriend/domain/coupon/dto/request/GrantMatchingCouponDto.java
@@ -1,9 +1,12 @@
 package com.hf.healthfriend.domain.coupon.dto.request;
 
+import com.hf.healthfriend.domain.coupon.entity.MatchingCoupon;
+import com.hf.healthfriend.domain.member.entity.Member;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import lombok.*;
 
+import java.time.Duration;
 import java.time.temporal.ChronoUnit;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -27,5 +30,15 @@ public class GrantMatchingCouponDto extends GrantCouponDto {
         super(receiverId, validTimeAmount, validTimeUnit);
         this.achievedLevel = achievedLevel;
         this.grantedMatchingCount = grantedMatchingCount;
+    }
+
+    @Override
+    public MatchingCoupon toEntity() {
+        return new MatchingCoupon(
+                new Member(super.getReceiverId()),
+                Duration.of(super.getValidTimeAmount(), super.getValidTimeUnit()),
+                this.achievedLevel,
+                this.grantedMatchingCount
+        );
     }
 }
