--- src/main/java/com/hf/healthfriend/domain/coupon/service/CouponServiceDelegator.java
@@ -1,6 +1,0 @@
-package com.hf.healthfriend.domain.coupon.service;
-
-public interface CouponServiceDelegator extends CouponService {
-
-    boolean supports(Class<?> dtoClass);
-}
